requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
gunicorn==20.1.0
python-dotenv==1.0.0
praw==7.7.1 # For Reddit API